// src/entities/enums.ts
export enum Role {
  GIAM_DOC = 'GIAM_DOC',
  TRUONG_PHONG = 'TRUONG_PHONG',
  TO_TRUONG = 'TO_TRUONG',
  CONG_NHAN = 'CONG_NHAN',
}

export enum ShiftName {
  C1 = 'C1',
  C2 = 'C2',
  C3 = 'C3',
}

export enum ShiftStatus {
  DRAFT = 'DRAFT',
  COMPLETED = 'COMPLETED',
  APPROVED = 'APPROVED',
  LOCKED = 'LOCKED',
}

export enum WorkOrderStatus {
  PENDING = 'PENDING',
  RUNNING = 'RUNNING',
  COMPLETED = 'COMPLETED',
}